# packages/openapi-spec/openapi.yaml
openapi: 3.0.3
info:
  title: NetPoint AI API
  description: API REST para el procesado de partidos de tenis, extracción de telemetría y consulta táctica multi-agente.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Servidor Local de Desarrollo
paths:
  /matches/process:
    post:
      summary: Iniciar procesamiento de partido (YouTube o Vídeo Local)
      description: Recibe la URL de un vídeo de YouTube o la ruta de un archivo para iniciar la ingesta y el tracking con visión por computador de forma asíncrona.
      operationId: processMatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessMatchRequest'
      responses:
        '202':
          description: Procesamiento aceptado e iniciado en segundo plano.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessMatchResponse'
        '400':
          $ref: '#/components/responses/400BadRequest'
...
components:
  schemas:
    ProcessMatchRequest:
      type: object
      additionalProperties: false
      required:
        - sourceUrl
      properties:
        sourceUrl:
          type: string
          format: uri
          minLength: 1
          example: "https://www.youtube.com/watch?v=dQw4w9WgXcQ"
        title:
          type: string
          example: "Final Alcaraz vs Sinner - Set 1"
    ProcessMatchResponse:
      type: object
      additionalProperties: false
      required:
        - matchId
        - status
        - message
      properties:
        matchId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        status:
          type: string
          enum: [processing]
          example: "processing"
        message:
          type: string
          example: "Procesamiento de vídeo iniciado en segundo plano."
...
  responses:
    400BadRequest:
      description: Solicitud incorrecta o parámetros faltantes.
...